REFACTOR: Move discrete (tabulator) doctrines to own submodules.
