+++ packages/catlog/src/dbl/discrete/mod.rs
@@ -1,0 +1,6 @@
+//! Doctrine of discrete double theories.
+
+pub mod model;
+pub mod model_diagram;
+pub mod model_morphism;
+pub mod theory;
