+++ packages/catlog/src/dbl/discrete_tabulator/mod.rs
@@ -1,0 +1,4 @@
+//! Doctrine of discrete tabulator theories.
+
+pub mod model;
+pub mod theory;
