+++ packages/catlog/src/dbl/mod.rs
@@ -8,3 +8,6 @@ pub mod model;
 pub mod model_diagram;
 pub mod model_morphism;
 pub mod theory;
+
+pub mod discrete;
+pub mod discrete_tabulator;
